project1/requirements.txt
Django==3.2.7
gunicorn==20.0.4
djangorestframework
django-cors-headers
aiohttp
httpx